Add rustic-compile as Control-O u
I overloaded control-O c in rust mode, but that is not sufficient, because I am
often in a Makefile or similar file that does not have the Rust mode. It's
unfortunate that rustic-compile is not just compile. In practice, however, the
usual compile deals very poorly with errors emitted by the Rust compiler